#94: Convert probabilities to floats internally, and in Lua. (.txt files still use integers, we will convert these to float as the files are converted one by one.)
git-svn-id: svn://www.knightsgame.org.uk/trunk@163 00979ad9-a3a6-4f7a-b49c-25ba5319b333

+++ src/engine/impl/action.hpp
@@ -242,7 +242,7 @@ public:
     virtual int getInt(int index) { error(); return 0; }
     virtual const ItemType * getItemType(int index) { error(); return 0; }
     virtual MapDirection getMapDirection(int index);  // default is to convert from a string
-    virtual int getProbability(int index) { error(); return 0; }
+    virtual float getProbability(int index) { error(); return 0; }
     virtual const KConfig::RandomInt * getRandomInt(int index) { error(); return 0; }
     virtual const Sound * getSound(int index) { error(); return 0; }
     virtual string getString(int index) { error(); return string(); }
